Authorizing

+++ application.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from flask import Flask, render_template
-from flask import request, redirect, jsonify, url_for
+from flask import request, redirect, jsonify, url_for,flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import User, City, Place, Base
@@ -235,10 +235,14 @@ def newCity():
 
 @app.route('/city/<int:city_id>/edit', methods=['GET', 'POST'])
 def editCity(city_id):
+
     if 'username' not in login_session:
         return redirect('/login')
-
+        
     editedCity = session.query(City).filter_by(id=city_id).one()
+    if editedCity.user_id != login_session['user_id']:
+        return render_template('authorize.html')
+
     if request.method == 'POST':
         if request.form['name']:
             editedCity.name = request.form['name']
@@ -255,6 +259,10 @@ def deleteCity(city_id):
         return redirect('/login')
 
     deletedCity = session.query(City).filter_by(id=city_id).one()
+
+    if deletedCity.user_id != login_session['user_id']:
+        return render_template('authorize.html')
+                    
     if request.method == 'POST':
         session.delete(deletedCity)
         session.commit()
@@ -300,8 +308,12 @@ def newPlace(city_id):
 def eidtPlace(city_id, place_id):
     if 'username' not in login_session:
         return redirect('/login')
-    city = session.query(City).filter_by(id=city_id).one()
+
     editedPlace = session.query(Place).filter_by(id=place_id).one()
+
+    if editedPlace.user_id != login_session['user_id']:
+        return render_template('authorize.html')
+
     if request.method == 'POST':
         if request.form['name']:
             editedPlace.name = request.form['name']
@@ -322,6 +334,10 @@ def deletePlace(city_id, place_id):
         return redirect('/login')
 
     deletedPlace = session.query(Place).filter_by(id=place_id).one()
+
+    if deletedPlace.user_id != login_session['user_id']:
+        return render_template('authorize.html')
+    
     if request.method == 'POST':
         session.delete(deletedPlace)
         session.commit()
